Changed use of make's shell to shell's backtick. Also fixed move command.
The 'if' for finding files in the 'done' silo is a shell command not make's so using $(shell) to get the output of 'find' is a mistake. Only backticks are required.
The move command to get cases from the 'done' silo is within the shell's 'if' statement so using '@' to silence it is erroneous and produces a shell error.